requirements.txt
flask==3.1.1
requests==2.32.3
apscheduler==3.11.0
python-dotenv==1.1.0
fpdf2==2.8.3
gunicorn==23.0.0
